+++ CMakeLists.txt
@@ -115,7 +115,6 @@ if (DEFINED ENV{BUILD_MCSEMA_DYNINST_DISASS})
     if($ENV{BUILD_MCSEMA_DYNINST_DISASS} EQUAL 1)
         add_custom_target(generate-protobuf-files DEPENDS ${PROJECT_PROTOBUFSOURCEFILES})
         add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/tools/mcsema_disass/dyninst)
-        add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/tools/mcsema_disass/dyninst_devel)
     endif()
 endif()
