+++ tools/mcsema_disass/dyninst/CFGWriter.cpp
@@ -17,42 +17,35 @@
 
 #include "Util.h"
 #include <cstdio>
+
 DECLARE_string(entrypoint);
+DECLARE_string(binary);
 
 DECLARE_bool(pie_mode);
+
 
 using namespace Dyninst;
 using namespace mcsema;
 
+mcsema::Module gModule;
+
 namespace {
 
-bool SmarterTryEval(InstructionAPI::Expression *expr,
+bool TryEval(InstructionAPI::Expression *expr,
                         const Address ip,
                         Address &result,
                         Address instruction_size=0) {
-  /*InstructionAPI::RegisterAST rip =
-      InstructionAPI::RegisterAST::makePC(Dyninst::Arch_x86_64);
-  auto ip_value = InstructionAPI::Result(InstructionAPI::u64, ip);
-  expr->bind(&rip, ip_value);
-*/
-  LOG(INFO) << expr->format();
-  /*auto res = expr->eval();
-  if (expr->eval().format() != "[empty]") {
-    LOG(INFO) << "Returned smart";
-    result = expr->eval().convert<Address>();
-    return true;
-  }*/
+
   if (auto bin = dynamic_cast<InstructionAPI::BinaryFunction *>(expr)) {
-    LOG(INFO) << "BF";
     std::vector<InstructionAPI::InstructionAST::Ptr> args;
     bin->getChildren(args);
 
     Address left, right;
 
-    auto first = SmarterTryEval(
+    auto first = TryEval(
         dynamic_cast<InstructionAPI::Expression *>(args[0].get()),
         ip, left, instruction_size);
-    auto second = SmarterTryEval(
+    auto second = TryEval(
         dynamic_cast<InstructionAPI::Expression *>(args[1].get()),
         ip, right, instruction_size);
     if (first && second) {
@@ -67,19 +60,16 @@ bool SmarterTryEval(InstructionAPI::Expression *expr,
       return false;
     }
   }  else if (auto imm = dynamic_cast<InstructionAPI::Immediate *>(expr)) {
-    LOG(INFO) << "IMM";
     result = imm->eval().convert<Address>();
     return true;
   } else if (auto deref = dynamic_cast<InstructionAPI::Dereference *>(expr)) {
-    LOG(INFO) << "DEREF";
     std::vector<InstructionAPI::InstructionAST::Ptr> args;
     deref->getChildren(args);
-    return SmarterTryEval(dynamic_cast<InstructionAPI::Expression *>(args[0].get()),
+    return TryEval(dynamic_cast<InstructionAPI::Expression *>(args[0].get()),
                    ip + instruction_size,
                    result);
   } else if (auto reg = dynamic_cast<InstructionAPI::RegisterAST *>(expr)) {
     if (reg->format() == "RIP") {
-      LOG(INFO) << "RIP " << ip << " + " << instruction_size;
       result = ip;
       return true;
     }
@@ -87,177 +77,160 @@ bool SmarterTryEval(InstructionAPI::Expression *expr,
   return false;
 }
 
+// Find call to __libc_start_main@plt and try to recover addresses from it
 Address TryRetrieveAddrFromStart(ParseAPI::CodeObject &code_object,
                                  Address start,
                                  size_t index) {
+  if (!start) {
+    LOG(WARNING) << "Binary starts at 0x0";
+    return 0;
+  }
   for (auto func : code_object.funcs()) {
     if (func->addr() == start) {
 
       auto entry_block = func->entry();
-      LOG(INFO) << "Start 0x" << std::hex << start
-                << " end 0x" << entry_block->end();
 
       using Insn = std::map<Offset, InstructionAPI::Instruction::Ptr>;
       Insn instructions;
       entry_block->getInsns(instructions);
-      //TODO(lukas): rename
-      auto callq = std::prev(instructions.end(), 2 + index);
+      auto call = std::prev(instructions.end(), 1);
 
-      LOG(INFO) << "Current rip 0x" << std::hex << callq->first;
+      // Just some sanity check that we are in correct function
+      if (call->second->getCategory() != InstructionAPI::c_CallInsn &&
+          call->second->getCategory() != InstructionAPI::c_BranchInsn) {
+        LOG(WARNING) << "Instruction at 0x" << std::hex << call->first
+                     << " is not call nor branching";
+        return 0;
+      }
 
-      auto second_operand = callq->second.get()->getOperand(1);
-      LOG(INFO) << second_operand.format(Arch_x86_64);
+      auto mov_inst = std::prev(call, 1 + index);
+      auto second_operand = mov_inst->second.get()->getOperand(1);
 
       Address offset = 0;
-      //lea
-      auto rip = callq->first;
-      if (callq->second->getOperation().getID() == 268) {
-        rip += callq->second->size();
+
+      // in -pie binaries it will be calculated using lea
+      auto rip = mov_inst->first;
+      if (mov_inst->second->getOperation().getID() == 268) {
+        rip += mov_inst->second->size();
       }
-      if (!SmarterTryEval(second_operand.getValue().get(), rip, offset, callq->second->size())) {
-        LOG(FATAL) << "Could not eval basic start addresses!";
+      if (!TryEval(second_operand.getValue().get(), rip, offset, mov_inst->second->size())) {
+        LOG(WARNING) << "Could not eval basic start addresses!";
+        return 0;
       }
       code_object.parse(offset, true);
       LOG(INFO) << "Retrieving info from _start at index " << index
                 << " got addr 0x" << std::hex << offset << std::dec;
       return offset;
-      //}
     }
   }
-  LOG(FATAL) << "Was not able to retrieve info from _start at index "
-             << index;
+  LOG(WARNING) << "Was not able to retrieve info from _start at index "
+               << index;
+  return 0;
 }
 
-bool IsInRegion(SymtabAPI::Region *r, Address a) {
-  if (a < r->getMemOffset()) {
-    return false;
-  }
-  if (a > (r->getMemOffset() + r->getMemSize())) {
-    return false;
-  }
-  return true;
-}
-
-bool IsInRegions(const std::vector<SymtabAPI::Region *> &regions, Address a) {
-  for (auto &r : regions) {
-    if (IsInRegion(r, a)) {
-      return true;
-    }
-  }
-  return false;
-}
-
-void WriteDataXref(const CFGWriter::CrossXref<mcsema::Segment> &xref,
+void WriteDataXref(const CrossXref<mcsema::Segment *> &xref,
                    const std::string &name="",
                    bool is_code=false,
                    uint64_t width=8) {
-  auto cfg_xref = xref.segment->add_xrefs();
-  cfg_xref->set_ea(xref.ea);
-  cfg_xref->set_width(width);
-  cfg_xref->set_target_ea(xref.target_ea);
-  cfg_xref->set_target_name(name);
-  cfg_xref->set_target_is_code(is_code);
-  cfg_xref->set_target_fixup_kind(mcsema::DataReference::Absolute);
+
+  auto cfg_xref = xref.WriteDataXref(name, is_code, width);
+  gDisassContext->data_xrefs.insert({static_cast<Dyninst::Address>(xref.ea), cfg_xref});
 }
 
-bool FishForXref(SymbolMap vars,
-                 const CFGWriter::CrossXref<mcsema::Segment> &xref,
-                 bool is_code=false, uint64_t width=8) {
-  auto var = vars.find(xref.target_ea);
-  if (var != vars.end()) {
-    WriteDataXref(xref, var->second, is_code, width);
-    return true;
+// Modifies gDisassContext
+void RenameFunc(Dyninst::Address ea, const std::string& new_name) {
+  LOG(INFO) << "Renaming 0x:" << std::hex << ea << " to " << new_name;
+  auto internal_func = gDisassContext->getInternalFunction(ea);
+  if (!internal_func) {
+    return;
   }
-  return false;
-}
-
-//TODO(lukas): Investigate, this ignores .bss?
-bool IsInBinary(ParseAPI::CodeSource &code_source, Address a) {
-  for (auto &r : code_source.regions()) {
-    if (r->contains(a)) {
-      return true;
-    }
-  }
-  LOG(INFO) << std::hex << a << std::dec << " is not contained in code_source";
-  return false;
+  internal_func->set_name(new_name);
 }
 
 } //namespace
 
-CFGWriter::CFGWriter(mcsema::Module &m, const std::string &module_name,
+CFGWriter::CFGWriter(mcsema::Module &m,
                      SymtabAPI::Symtab &symtab,
                      ParseAPI::SymtabCodeSource &symCodeSrc,
-                     ParseAPI::CodeObject &codeObj,
-                     const ExternalFunctionManager &extFuncMgr,
-                     Address entry_point)
+                     ParseAPI::CodeObject &codeObj)
     : module(m),
-      module_name(module_name),
       symtab(symtab),
       code_source(symCodeSrc),
       code_object(codeObj),
-      ext_func_manager(extFuncMgr),
-      entry_point(entry_point) {
+      magic_section(gDisassContext->magic_section),
+      ptr_byte_size(symtab.getAddressWidth()){
 
-  // Populate skip_funcss with some functions known to cause problems
-  skip_funcss = {
-                 "_fini",
-                 "__libc_start_main"
-                 };
-
-  // Populate func_map
-
-  std::vector<SymtabAPI::Function *> functions;
-  symtab.getAllFunctions(functions);
-  bool is_stripped = (functions.size()) ? false : true;
-  LOG(INFO) << "Binary is stripped: " << is_stripped;
+  LOG(INFO) << "Binary is stripped: " << symtab.isStripped();
 
   std::vector<SymtabAPI::Region *> regions;
   symtab.getAllRegions(regions);
 
   for (auto reg : regions) {
-      section_manager.AddRegion(reg);
+      gSectionManager->AddRegion(reg);
+      if (reg->getMemOffset()) {
+        gDisassContext->segment_eas.push_back(reg->getMemOffset());
+      }
   }
+
 
   // We need to get main! Heuristic for stripped binaries is that main is
   // passed to __libc_start_main as last argument from _start, which we can
   // find, because it is entrypoint
-  code_object.parse(entry_point, true);
+
+  // This does NOT return 0 for shared libraries
+  Address entry_point = symtab.getEntryOffset();
+
+  LOG(INFO) << "Entry offset is 0x" << std::hex << entry_point;
+  if (entry_point) {
+    code_object.parse(entry_point, true);
+  }
   Address main_offset = TryRetrieveAddrFromStart(code_object, entry_point, 0);
   Address ctor_offset = TryRetrieveAddrFromStart(code_object, entry_point, 1);
   Address dtor_offset = TryRetrieveAddrFromStart(code_object, entry_point, 2);
 
-  // TODO(lukas): Check if there's a better way
+  LOG_IF(WARNING, main_offset) << "Entrypoint was not found!";
 
-  CHECK(main_offset) << "Entrypoint with name "
-                     << FLAGS_entrypoint
-                     <<" was not found!";
-
+  // TODO(lukas): When lifting shared library internal functions that needs
+  // to be entrypoints are demangled by ParseAPI, which is really unfortunate
   for (auto func : code_object.funcs()) {
-    func_map[func->addr()] = func->name();
-    LOG(INFO) << func->addr() << " " << func->name();
-  }
-  func_map[main_offset] = FLAGS_entrypoint;
-  LOG(INFO) << "Function at " << main_offset << " is " << FLAGS_entrypoint;
+    auto cfg_internal_func = module.add_funcs();
 
-  if (func_map[ctor_offset].substr(0, 4) == "targ") {
-    LOG(INFO) << "Renaming 0x:" << std::hex << ctor_offset << " to init";
-    func_map[ctor_offset] = "init";
-  }
-
-  if (func_map[dtor_offset].substr(0, 4) == "targ") {
-    LOG(INFO) << "Renaming 0x:" << std::hex << dtor_offset << " to fini";
-    func_map[dtor_offset] = "fini";
-  }
-
-
-  for (auto reg : regions) {
-    if (reg->getRegionName() == ".text") {
-      relocations = reg->getRelocations();
+    SymtabAPI::Function *symtab_func;
+    if (symtab.findFuncByEntryOffset(symtab_func, func->addr())) {
+      LOG(INFO) << "Taking mangled name from symtab";
+      cfg_internal_func->set_name(*(symtab_func->mangled_names_begin()));
+      auto beg = symtab_func->mangled_names_begin();
+      while (beg != symtab_func->mangled_names_end()) {
+        LOG(INFO) << "\t" << *beg;
+        ++beg;
+      }
+    } else {
+      cfg_internal_func->set_name(func->name());
     }
+    cfg_internal_func->set_ea(func->addr());
+    cfg_internal_func->set_is_entrypoint(false);
+    gDisassContext->func_map.insert({func->addr(), cfg_internal_func});
+    LOG(INFO) << "Found internal function at 0x" << func->addr()
+              << " with name " << cfg_internal_func->name();
   }
 
-  getNoReturns();
+  // give entrypoint correct name, most likely main
+  if (main_offset) {
+    RenameFunc(main_offset, FLAGS_entrypoint);
+  }
 
+  // We need to give libc ctor/dtor names
+  if (ctor_offset) {
+    RenameFunc(ctor_offset, "init");
+  }
+  if (dtor_offset) {
+    RenameFunc(dtor_offset, "fini");
+  }
+
+  GetNoReturns();
+
+  // Calculate where can magic section start without
+  // potentialy overwriting part of the binary
   //TODO(lukas): Move out
   Address highest = 0;
   for (auto reg : regions) {
@@ -265,46 +238,52 @@ CFGWriter::CFGWriter(mcsema::Module &m, const std::string &module_name,
   }
   highest += 0x420;
   LOG(INFO) << "Magic section starts at 0x" << std::hex << highest;
-  magic_section.start_ea = highest;
+  magic_section.init(highest, ptr_byte_size);
 }
 
-void CFGWriter::write() {
-  writeExternalFunctions();
-  writeExternalVariables();
-  LOG(INFO) << "Writing internal data";
-  writeInternalData();
-  writeGlobalVariables();
+void CFGWriter::WriteFunction(Dyninst::ParseAPI::Function *func,
+                              mcsema::Function *cfg_internal_func) {
+  gDisassContext->func_map.insert({func->addr(), cfg_internal_func});
 
-  writeInternalFunctions();
+  ParseAPI::Block *entryBlock = func->entry();
+  cfg_internal_func->set_ea(entryBlock->start());
+
+  cfg_internal_func->set_is_entrypoint(func);
+
+  for (ParseAPI::Block *block : func->blocks()) {
+    WriteBlock(block, func, cfg_internal_func);
+  }
+
+  cfg_internal_func->set_name(func->name());
+  LOG(INFO) << "Added " << func->name() << " into module, found via xref";
+}
+
+void CFGWriter::Write() {
+  WriteExternalFunctions();
+  WriteExternalVariables();
+  WriteInternalData();
+  WriteGlobalVariables();
+
+  WriteInternalFunctions();
 
   //Handle new functions found via various xrefs, mostly in stripped binary
   LOG(INFO) << code_xrefs_to_resolve.size() << " code xrefs is unresolved!";
   for (auto &a : code_xrefs_to_resolve) {
-    code_object.parse(a.first, true);
     WriteDataXref(a.second, "", true);
+    code_object.parse(a.first, true);
   }
 
   for (auto func : code_object.funcs()) {
-    if (code_xrefs_to_resolve.find(func->addr()) != code_xrefs_to_resolve.end() &&
-        func_map.find(func->addr()) == func_map.end()) {
+    auto code_xref = code_xrefs_to_resolve.find(func->addr());
+    if (code_xref != code_xrefs_to_resolve.end() &&
+        !gDisassContext->getInternalFunction(func->addr())) {
 
-      func_map[func->addr()] = func->name();
-      auto cfgInternalFunc = module.add_funcs();
-
-      ParseAPI::Block *entryBlock = func->entry();
-      cfgInternalFunc->set_ea(entryBlock->start());
-
-      cfgInternalFunc->set_is_entrypoint(func);
-
-      for (ParseAPI::Block *block : func->blocks()) {
-        writeBlock(block, func, cfgInternalFunc);
-      }
-
-      cfgInternalFunc->set_name(func_map[func->addr()]);
-      LOG(INFO) << "Added " << func->name() << " into module, found via xref";
+      auto cfg_internal_func = module.add_funcs();
+      WriteFunction(func, cfg_internal_func);
     }
   }
 
+  // In case we discovered some new ones, we need to try until fixpoint
   while (!inst_xrefs_to_resolve.empty()) {
     LOG(WARNING) << inst_xrefs_to_resolve.size() << " inst code xrefs is unresolved!";
     for (auto &a : inst_xrefs_to_resolve) {
@@ -315,103 +294,70 @@ void CFGWriter::write() {
     inst_xrefs_to_resolve.clear();
     for (auto func : code_object.funcs()) {
       if (old_set.find(func->addr()) != old_set.end() &&
-          func_map.find(func->addr()) == func_map.end()) {
+          !gDisassContext->getInternalFunction(func->addr())) {
 
-        func_map[func->addr()] = func->name();
-        auto cfgInternalFunc = module.add_funcs();
-
-        ParseAPI::Block *entryBlock = func->entry();
-        cfgInternalFunc->set_ea(entryBlock->start());
-
-        cfgInternalFunc->set_is_entrypoint(func);
-
-        for (ParseAPI::Block *block : func->blocks()) {
-          writeBlock(block, func, cfgInternalFunc);
-        }
-
-        cfgInternalFunc->set_name(func_map[func->addr()]);
-        LOG(INFO) << "Added " << func->name() << " into module, found via xref";
+        auto cfg_internal_func = module.add_funcs();
+        WriteFunction(func, cfg_internal_func);
       }
     }
   }
-  module.set_name(module_name);
+  module.set_name(FLAGS_binary);
 }
 
-void CFGWriter::writeExternalVariables() {
+void CFGWriter::WriteExternalVariables() {
   std::vector<SymtabAPI::Symbol *> symbols;
   symtab.getAllSymbolsByType(symbols, SymtabAPI::Symbol::ST_OBJECT);
 
   LOG(INFO) << "Writing " << symbols.size() << " external variables";
   for (const auto &s : symbols) {
     if (s->isInDynSymtab()) {
-      // Most likely relocation. So as IDA we can also make up some addr
+      // Most likely relocation. So as IDA we can also make up some addresses
       if (!s->getOffset() && !s->getSize()) {
-        LOG(WARNING)
-            << "External var has no ea! " << s->getMangledName();
-        LOG(WARNING)
-            << "External var has no size! " << s->getMangledName();
-
-        Address unreal_ea = magic_section.start_ea + magic_section.size;
-        auto extVar = module.add_external_vars();
-        extVar->set_name(s->getMangledName());
-        extVar->set_ea(unreal_ea);
-
-        extVar->set_size(ptr_byte_size);
-
-        //TODO(lukas): This needs some checks
-        extVar->set_is_weak(false);
-        extVar->set_is_thread_local(false);
-
-        magic_section.size += ptr_byte_size;
-        for (int i = 0; i < ptr_byte_size; ++i) {
-          magic_section.data << "\0";
-        }
-        external_vars.insert( { unreal_ea, s->getMangledName() } );
-        magic_section.ext_vars.push_back(extVar);
+        LOG(INFO) << "External var " << s->getMangledName()
+                  << " had no ea, allocating it in magic section";
+        auto external_var = magic_section.WriteExternalVariable(
+            module, s->getMangledName());
+        gDisassContext->external_vars.insert({external_var->ea(), external_var});
         continue;
       }
-      external_vars.insert( { s->getOffset(), s->getMangledName() } );
 
-      auto extVar = module.add_external_vars();
-      extVar->set_name(s->getMangledName());
-      extVar->set_ea(s->getOffset());
+      auto external_var = module.add_external_vars();
+      gDisassContext->external_vars.insert({s->getOffset(), external_var});
+      external_var->set_name(s->getMangledName());
+      external_var->set_ea(s->getOffset());
 
       //TODO(lukas): This is to generate syntactically valid llvm
       //             Other than that probably won't work
       if (s->getSize()) {
-        extVar->set_size(s->getSize());
+        external_var->set_size(s->getSize());
       } else {
-        extVar->set_size(8);
+        external_var->set_size(ptr_byte_size);
       }
 
       //TODO(lukas): This needs some checks
-      extVar->set_is_weak(false);
-      extVar->set_is_thread_local(false);
+      external_var->set_is_weak(false);
+      external_var->set_is_thread_local(false);
     }
   }
 }
 
-void CFGWriter::writeGlobalVariables() {
+void CFGWriter::WriteGlobalVariables() {
   std::vector<SymtabAPI::Symbol *> vars;
   for (auto &a : vars) {
     if (a->getRegion() && (a->getRegion()->getRegionName() == ".bss" ||
                            a->getRegion()->getRegionName() == ".rodata")) {
       LOG(INFO) << "Found global variable " << a->getMangledName()
                 << " at " << std::hex << a->getOffset() << std::dec;
-      auto globalVar = module.add_global_vars();
-      globalVar->set_ea(a->getOffset());
-      globalVar->set_name(a->getMangledName());
-      globalVar->set_size(a->getSize());
-      global_vars.insert({a->getOffset(), a->getMangledName()});
+      auto global_var = module.add_global_vars();
+      global_var->set_ea(a->getOffset());
+      global_var->set_name(a->getMangledName());
+      global_var->set_size(a->getSize());
+      gDisassContext->global_vars.insert({a->getOffset(), global_var});
     }
   }
 }
 
-bool CFGWriter::shouldSkipFunction(const std::string &name) const {
-  return skip_funcss.find(name) != skip_funcss.end();
-}
-
-void CFGWriter::writeInternalFunctions() {
+void CFGWriter::WriteInternalFunctions() {
     // I don't want this to be in recompiled binary as compiler will
     // add them as well, sub_* is enough
     std::unordered_set< std::string > notEntryPoints = {
@@ -431,42 +377,44 @@ void CFGWriter::writeInternalFunctions() {
     };
 
   for (ParseAPI::Function *func : code_object.funcs()) {
-    if (shouldSkipFunction(func->name()))
-      continue;
-    else if (isExternal(func->entry()->start()))
-      continue;
+   if (IsExternal(func->entry()->start())) {
+     LOG(INFO) << "Function " << func->name() << " is getting skipped";
+     continue;
+   }
     // We want to ignore the .got.plt stubs, since they are not needed
     // and cfg file would grow significantly
-    else if (IsInRegion(section_manager.getRegion(".got.plt"),
-                        func->entry()->start())) {
+    else if (gSectionManager->IsInRegion(".got.plt",
+                                          func->entry()->start())) {
+     LOG(INFO) << "Function " << func->name()
+               << " is getting skipped because it is .got.pl stub";
       continue;
     }
 
-    auto cfgInternalFunc = module.add_funcs();
+    auto cfg_internal_func = gDisassContext->getInternalFunction(func->addr());
 
     ParseAPI::Block *entryBlock = func->entry();
-    cfgInternalFunc->set_ea(entryBlock->start());
+    CHECK(entryBlock->start() == cfg_internal_func->ea())
+        << "Start of the block is not equal to function ea";
 
-    cfgInternalFunc->set_is_entrypoint(
+    cfg_internal_func->set_is_entrypoint(
         notEntryPoints.find(func->name()) == notEntryPoints.end());
+    LOG(INFO) << "Function " << cfg_internal_func->name() << " at 0x"
+              << std::hex << cfg_internal_func->ea()
+              << " is entry point? "<< cfg_internal_func->is_entrypoint();
 
     for (ParseAPI::Block *block : func->blocks()) {
-      writeBlock(block, func, cfgInternalFunc);
-    }
-
-    if (func_map.find(func->addr()) != func_map.end()) {
-      cfgInternalFunc->set_name(func_map[func->addr()]);
+      WriteBlock(block, func, cfg_internal_func);
     }
   }
 }
 
 
 //TODO(lukas): This one is basically unchanged from original PR.
-void CFGWriter::writeBlock(ParseAPI::Block *block, ParseAPI::Function *func,
-                           mcsema::Function *cfgInternalFunc) {
+void CFGWriter::WriteBlock(ParseAPI::Block *block, ParseAPI::Function *func,
+                           mcsema::Function *cfg_internal_func) {
 
-  mcsema::Block *cfgBlock = cfgInternalFunc->add_blocks();
-  cfgBlock->set_ea(block->start());
+  mcsema::Block *cfg_block = cfg_internal_func->add_blocks();
+  cfg_block->set_ea(block->start());
 
   // Set outgoing edges
   for (auto edge : block->targets()) {
@@ -501,7 +449,7 @@ void CFGWriter::writeBlock(ParseAPI::Block *block, ParseAPI::Function *func,
     }
 
     if (!found)
-      cfgBlock->add_successor_eas(edge->trg()->start());
+      cfg_block->add_successor_eas(edge->trg()->start());
   }
 
   // Write instructions
@@ -514,77 +462,47 @@ void CFGWriter::writeBlock(ParseAPI::Block *block, ParseAPI::Function *func,
   for (auto p : instructions) {
     InstructionAPI::Instruction *instruction = p.second.get();
 
-    writeInstruction(instruction, ip, cfgBlock);
+    WriteInstruction(instruction, ip, cfg_block);
     ip += instruction->size();
   }
 
 }
 
-void CFGWriter::writeInstruction(InstructionAPI::Instruction *instruction,
-                                 Address addr, mcsema::Block *cfgBlock) {
+void CFGWriter::WriteInstruction(InstructionAPI::Instruction *instruction,
+                                 Address addr, mcsema::Block *cfg_block) {
 
-  mcsema::Instruction *cfgInstruction = cfgBlock->add_instructions();
+  mcsema::Instruction *cfg_instruction = cfg_block->add_instructions();
 
   std::string instBytes;
   for (int offset = 0; offset < instruction->size(); ++offset) {
     instBytes += (int)instruction->rawByte(offset);
   }
 
-  cfgInstruction->set_bytes(instBytes);
-  cfgInstruction->set_ea(addr);
+  cfg_instruction->set_bytes(instBytes);
+  cfg_instruction->set_ea(addr);
 
   std::vector<InstructionAPI::Operand> operands;
   instruction->getOperands(operands);
 
   if (instruction->getCategory() == InstructionAPI::c_CallInsn) {
-    handleCallInstruction(instruction, addr, cfgInstruction);
+    HandleCallInstruction(instruction, addr, cfg_instruction);
   } else {
-    handleNonCallInstruction(instruction, addr, cfgInstruction);
+    HandleNonCallInstruction(instruction, addr, cfg_instruction);
   }
   //TODO(lukas): We just found it, no need to bother with it?
   //             But what if it needs to be start of function?
   code_xrefs_to_resolve.erase(addr);
 }
 
-/* DEPRECATED
-bool getDisplacement(InstructionAPI::Instruction *instruction,
-                     std::set<InstructionAPI::Expression::Ptr> &operands,
-                     Address &address) {
-  bool found = false;
 
-  LOG(INFO) << "Checking for displacement at " << instruction->format() << " " << operands.size();
-  for (auto &op : operands) {
-    if (auto binFunc =
-            dynamic_cast<InstructionAPI::BinaryFunction *>(op.get())) {
-
-      std::vector<InstructionAPI::InstructionAST::Ptr> operands;
-      binFunc->getChildren(operands);
-
-      for (auto &a : operands) {
-        if (auto second =
-                dynamic_cast<InstructionAPI::BinaryFunction *>(a.get()))
-          found = true;
-      }
-
-      for (auto &a : operands) {
-        if (auto addr =
-                dynamic_cast<InstructionAPI::Immediate *>(operands[0].get())) {
-          address = addr->eval().convert<Address>();
-        }
-      }
-    }
-  }
-  return found && address;
-}
-*/
-void writeDisplacement(mcsema::Instruction *cfgInstruction, Address &address,
+void WriteDisplacement(mcsema::Instruction *cfg_instruction, Address &address,
         const std::string& name = "") {
   // Addres is uint64_t and in CFG ea is int64_t
   if (static_cast<int64_t>(address) <= 0) {
     return;
   }
 
-  AddCodeXref( cfgInstruction, CodeReference::DataTarget,
+  AddCodeXref( cfg_instruction, CodeReference::DataTarget,
           CodeReference::MemoryDisplacementOperand, CodeReference::Internal,
           address, name );
 }
@@ -605,12 +523,13 @@ Address DisplacementHelper(Dyninst::InstructionAPI::Expression *expr) {
   return 0;
 }
 
-void CFGWriter::checkDisplacement(Dyninst::InstructionAPI::Expression *expr,
-                       mcsema::Instruction *cfgInstruction) {
+// TODO(lukas): It has to actually point at something to be MDO
+void CFGWriter::CheckDisplacement(Dyninst::InstructionAPI::Expression *expr,
+                       mcsema::Instruction *cfg_instruction) {
 
   //TODO(lukas): This is possibly incorrect attempt to cull down amount of
   //             "false" xrefs of type MemoryDisplacement
-  if (cfgInstruction->xrefs_size()) {
+  if (cfg_instruction->xrefs_size()) {
     LOG(INFO) << "Avoiding clobbering with xrefs!";
     return;
   }
@@ -622,18 +541,18 @@ void CFGWriter::checkDisplacement(Dyninst::InstructionAPI::Expression *expr,
       if (auto inner_expr = dynamic_cast<InstructionAPI::Expression *>(op.get())) {
         auto displacement = DisplacementHelper(inner_expr);
         if (displacement) {
-          writeDisplacement(cfgInstruction, displacement);
+          WriteDisplacement(cfg_instruction, displacement);
         }
       }
     }
   } else {
     if (auto displacement = DisplacementHelper(expr)) {
-      writeDisplacement(cfgInstruction, displacement);
+      WriteDisplacement(cfg_instruction, displacement);
     }
   }
 }
 
-void CFGWriter::getNoReturns() {
+void CFGWriter::GetNoReturns() {
   for (auto f : code_object.funcs()) {
     if (f->retstatus() == ParseAPI::NORETURN) {
       no_ret_funcs.insert(f->name());
@@ -641,43 +560,15 @@ void CFGWriter::getNoReturns() {
   }
 }
 
-bool CFGWriter::isNoReturn(const std::string &name) {
+bool CFGWriter::IsNoReturn(const std::string &name) {
   return no_ret_funcs.find(name) != no_ret_funcs.end();
 }
 
-//TODO(lukas): This one should be replaced in favor of
-//             function of the type handleXref or FishFor*
-std::string CFGWriter::getXrefName(Address addr) {
-  auto func = func_map.find(addr);
-  if (func != func_map.end()) {
-    return func->second;
-  }
 
-  auto extVar = external_vars.find(addr);
-  if (extVar != external_vars.end()) {
-    return extVar->second;
-  }
-
-  auto globalVar = global_vars.find(addr);
-  if (globalVar != global_vars.end()) {
-    return globalVar->second;
-  }
-
-
-  auto segmentVar = segment_vars.find(addr);
-  if (segmentVar != segment_vars.end()) {
-    return segmentVar->second;
-  }
-
-  // Sometimes things have no name but they are still xrefs
-  // We could leave the name empty as well
-  return "__mcsema_unknown";
-}
-
-//TODO(lukas): This is hacky af
-void CFGWriter::handleCallInstruction(InstructionAPI::Instruction *instruction,
+//TODO(lukas): This is hacky
+void CFGWriter::HandleCallInstruction(InstructionAPI::Instruction *instruction,
                                       Address addr,
-                                      mcsema::Instruction *cfgInstruction) {
+                                      mcsema::Instruction *cfg_instruction) {
   Address target;
   std::vector<InstructionAPI::Operand> operands;
   instruction->getOperands(operands);
@@ -686,241 +577,97 @@ void CFGWriter::handleCallInstruction(InstructionAPI::Instruction *instruction,
 
   LOG(INFO) << "TryEval " << addr << " " << size;
   bool got_result = false;
-  got_result = SmarterTryEval(operands[0].getValue().get(), addr, target, size);
+  if (TryEval(operands[0].getValue().get(), addr, target, size)) {
+    HandleXref(cfg_instruction, target);
 
-  if (!got_result) {
-    got_result = tryEval(operands[0].getValue().get(), addr + size, target);
-    target -= size;
-    //TODO(lukas): Move this out on release!
-    LOG(INFO) << "Need old " << got_result;
-    CHECK(!got_result) << "Truly needed old one!";
-  }
-  if (got_result) {
-    handleXref(cfgInstruction, target);
-    // What can happen is that we get xref somewhere in the .text and handleXref
+    // What can happen is that we get xref somewhere in the .text and HandleXref
     // fills it with defaults. We need to check and correct it if needed
-    if (IsInRegion(section_manager.getRegion(".text"), target)) {
-      auto xref = cfgInstruction->mutable_xrefs(0);
-      xref->set_target_type(CodeReference::CodeTarget);
+    if (gSectionManager->IsInRegion(".text", target)) {
+      auto xref = cfg_instruction->mutable_xrefs(0);
+      //xref->set_target_type(CodeReference::CodeTarget);
       xref->set_operand_type(CodeReference::ControlFlowOperand);
 
       // It is pointing in .text and not to a function?
       // That's weird, quite possibly we are missing a function!
-      if (func_map.find(target) == func_map.end()) {
+      if (!gDisassContext->getInternalFunction(target)) {
         LOG(INFO) << "Unresolved inst_xref " << target;
         inst_xrefs_to_resolve.insert(
-          {target , {addr, target, cfgInstruction}});
+          {target , {addr, target, cfg_instruction}});
       }
     }
 
-    if (isNoReturn(getExternalName(target))) {
-      cfgInstruction->set_local_noreturn(true);
+    if (IsNoReturn(cfg_instruction->mutable_xrefs(0)->name())) {
+      cfg_instruction->set_local_noreturn(true);
     }
     return;
   }
-
-
-  // Below is territory still occupied by old PR and I am not sure what to do
-  // with it.
-  // It didn't work without it, but it does now, magic?
-  // Keeping it to see how relocs were used
-  LOG(INFO) << "CallInst fallback to default " << addr;
-/*
-  if (relocations.size() > 0) {
-    auto entry = *(relocations.begin());
-    bool found = false;
-
-    for (auto ent : relocations) {
-      if (ent.rel_addr() == (addr + size + 1)) {
-        entry = ent;
-        found = true;
-        break;
-      }
-    }
-
-    if (!((!found) || (found && !entry.getDynSym()->getRegion()))) {
-        Offset off = entry.getDynSym()->getOffset();
-
-        AddCodeXref(cfgInstruction,
-                    CodeReference::CodeTarget,
-                    CodeReference::ControlFlowOperand,
-                    CodeReference::Internal,
-                    off);
-      return;
-    }
-  }
-*/
-  // TODO: This is when things like callq %*rcx happen, not sure yet how to deal
-  // with it correctly, this seems to work good enough for now
-
-  auto name = getXrefName(target);
-  if (name == "__mcsema_unknown") {
-    return;
-  }
-  AddCodeXref(cfgInstruction,
-              CodeReference::CodeTarget,
-              CodeReference::ControlFlowOperand,
-              CodeReference::Internal,
-              target, name);
 }
 
 Address CFGWriter::immediateNonCall( InstructionAPI::Immediate* imm,
-        Address addr, mcsema::Instruction* cfgInstruction ) {
+        Address addr, mcsema::Instruction* cfg_instruction ) {
 
   Address a = imm->eval().convert<Address>();
-  auto allSymsAtOffset = symtab.findSymbolByOffset(a);
-  bool isRef = false;
-
-  if (allSymsAtOffset.size() > 0) {
-    for (auto symbol : allSymsAtOffset) {
-      if (symbol->getType() == SymtabAPI::Symbol::ST_OBJECT) {
-        isRef = true;
-      }
-    }
-  }
-
-  // TODO(lukas): This is ugly hack from previous PR, but it works for now
-  // Almost certainly needs rework because of -fPIC and -pie
-  if (a > 0x10000) {
-    isRef = true;
-  }
-  if (a > 0xffffffffffffff00) {
-    isRef = false;
-  }
-
-  if (isRef) {
-    auto cfgCodeRef = AddCodeXref(cfgInstruction,
-                                CodeReference::DataTarget,
-                                CodeReference::ImmediateOperand,
-                                CodeReference::Internal,
-                                a, getXrefName(a));
-
-    if (isExternal(a) ||
-        external_vars.find(a) != external_vars.end()) {
-        cfgCodeRef->set_location(CodeReference::External);
-    }
-    if (getXrefName(a) == "__mcsema_unknown" &&
-        IsInRegion(section_manager.getRegion(".text"), a)) {
+  if (!gDisassContext->HandleCodeXref({addr, a, cfg_instruction}, false)) {
+    if (gSectionManager->IsInRegion(".text", a)) {
+      auto cfgCodeRef = AddCodeXref(cfg_instruction,
+                            CodeReference::DataTarget,
+                            CodeReference::ImmediateOperand,
+                            CodeReference::Internal,
+                            a);
       LOG(INFO) << std::hex
                 << "IMM may be working with new function starting at" << a;
       inst_xrefs_to_resolve.insert({a, {}});
+      return a;
     }
-    return a;
+    LOG(INFO) << "Not forcing target 0x" << std::hex << a;
+    return 0;
   }
-  return 0;
+  auto cfg_code_xref = cfg_instruction->mutable_xrefs(cfg_instruction->xrefs_size() - 1);
+  cfg_code_xref->set_operand_type(CodeReference::ImmediateOperand);
+  return a;
+
 }
 
 Address CFGWriter::dereferenceNonCall(InstructionAPI::Dereference* deref,
                                    Address addr,
-                                   mcsema::Instruction* cfgInstruction) {
+                                   mcsema::Instruction* cfg_instruction) {
 
   std::vector<InstructionAPI::InstructionAST::Ptr> children;
   deref->getChildren(children);
   auto expr = dynamic_cast<InstructionAPI::Expression *>(children[0].get());
 
-  if (!expr) {
-    throw std::runtime_error{"expected expression"};
-  }
+  CHECK(expr) << "Expected expression";
 
   Address a;
-  if (tryEval(expr, addr, a)) {
-    //TODO(lukas): Possibly wrong with -pie & -fPIC?
-    if (a > 0xffffffffffffff00) {
-      return 0;
-    }
-
-    auto cfgCodeRef = AddCodeXref(cfgInstruction,
-                                  CodeReference::DataTarget,
-                                  CodeReference::MemoryOperand,
-                                  CodeReference::Internal, a,
-                                  getXrefName(a));
-    if (isExternal(a) || external_vars.find(a) != external_vars.end()) {
-      cfgCodeRef->set_location(CodeReference::External);
-    }
+  // TODO(lukas): Possibly may discover new functions?
+  if (TryEval(expr, addr, a)) {
+    gDisassContext->HandleCodeXref({addr, a, cfg_instruction});
     return a;
   }
 
   return 0;
 }
 
+// Handling only CODEXREFS!
+// TODO(lukas): gDisassContext should be responsible for this
+bool CFGWriter::HandleXref(mcsema::Instruction *cfg_instruction,
+                           Address addr,
+                           bool force) {
 
-void CFGWriter::handleXref(mcsema::Instruction *cfg_instruction,
-                           Address addr) {
-  LOG(INFO) << "HandleXref to " << addr;
-  auto func = func_map.find(addr);
-  if (func != func_map.end()) {
-    LOG(INFO) << "Code xref to func at " << addr;
-    auto code_ref = AddCodeXref(cfg_instruction,
-                                  CodeReference::CodeTarget,
-                                  CodeReference::ControlFlowOperand,
-                                  CodeReference::Internal, addr,
-                                  func->second);
-    if (isExternal(addr)) {
-      code_ref->set_location(CodeReference::External);
-    }
-    return;
-  }
-
-  auto g_var = global_vars.find(addr);
-  if (g_var != global_vars.end()) {
-    auto code_ref = AddCodeXref(cfg_instruction,
-                                  CodeReference::DataTarget,
-                                  CodeReference::MemoryOperand,
-                                  CodeReference::Internal, addr,
-                                  g_var->second);
-    return;
-  }
-
-  auto s_var = segment_vars.find(addr);
-  if (s_var != segment_vars.end()) {
-    auto code_ref = AddCodeXref(cfg_instruction,
-                                  CodeReference::DataTarget,
-                                  CodeReference::MemoryOperand,
-                                  CodeReference::Internal, addr,
-                                  s_var->second);
-    return;
-  }
-
-  auto ext_var = external_vars.find(addr);
-  if (ext_var != external_vars.end()) {
-    LOG(INFO) << "Code xref to ext_var at " << addr;
-    auto code_ref = AddCodeXref(cfg_instruction,
-                                  CodeReference::DataTarget,
-                                  CodeReference::MemoryOperand,
-                                  CodeReference::External, addr,
-                                  ext_var->second);
-    return;
-  }
-  auto ext_func = external_functions.find(addr);
-  if (ext_func != external_functions.end()) {
-    LOG(INFO) << "Code xref to ext_func at " << addr;
-    auto code_ref = AddCodeXref(cfg_instruction,
-                                  CodeReference::CodeTarget,
-                                  CodeReference::ControlFlowOperand,
-                                  CodeReference::External, addr,
-                                  ext_func->second);
-    return;
-  }
-  LOG(INFO) << "Could not recognize xref anywhere falling to default";
-  auto code_ref = AddCodeXref(cfg_instruction,
-                                CodeReference::DataTarget,
-                                CodeReference::MemoryOperand,
-                                CodeReference::Internal, addr);
-  if (isExternal(addr)) {
-    code_ref->set_location(CodeReference::External);
-  }
-
+  return gDisassContext->HandleCodeXref({0, addr, cfg_instruction}, force);
 }
 
-void CFGWriter::handleNonCallInstruction(
+void CFGWriter::HandleNonCallInstruction(
     Dyninst::InstructionAPI::Instruction *instruction,
     Address addr,
-    mcsema::Instruction *cfgInstruction) {
+    mcsema::Instruction *cfg_instruction) {
 
   std::vector<InstructionAPI::Operand> operands;
   instruction->getOperands(operands);
 
   // RIP already points to the next instruction
+  // Except sometimes DynInst thinks it doesn't
+  // and construct an AST with + |instruction|
   addr += instruction->size();
 
   // Sometimes some .text address is stored somewhere in data segment.
@@ -931,83 +678,58 @@ void CFGWriter::handleNonCallInstruction(
   LOG(INFO) << instruction->format() << " at 0x" << std::hex << addr - instruction->size();
   for (auto op : operands) {
     auto expr = op.getValue();
-    LOG(INFO) << expr->format();
     if (auto imm = dynamic_cast<InstructionAPI::Immediate *>(expr.get())) {
-      LOG(INFO) << "Dealing with Immidiate as op at 0x" << std::hex << addr - instruction->size();
-      direct_values[i] = immediateNonCall( imm, addr, cfgInstruction);
+      LOG(INFO) << "\tDealing with Immidiate as op";
+      direct_values[i] = immediateNonCall(imm, addr, cfg_instruction);
     } else if (
         auto deref = dynamic_cast<InstructionAPI::Dereference *>(expr.get())) {
-      LOG(INFO) << "Dealing with Dereference as op at 0x" << std::hex << addr - instruction->size();
-      direct_values[i] = dereferenceNonCall( deref, addr, cfgInstruction);
+      LOG(INFO) << "\tDealing with Dereference as op";
+      direct_values[i] = dereferenceNonCall(deref, addr, cfg_instruction);
     } else if (
         auto bf = dynamic_cast<InstructionAPI::BinaryFunction *>(expr.get())) {
-      LOG(INFO) << "Dealing with BinaryFunction as op at 0x" << std::hex << addr - instruction->size();
+      LOG(INFO) << "Dealing with BinaryFunction as op";
 
       // 268 stands for lea
-      // 8 stands for jump
       auto instruction_id = instruction->getOperation().getID();
       if (instruction_id == 268) {
         Address a;
-        /*if (addr - instruction->size() == 0x405c7d) {
-          auto xref = AddCodeXref(cfgInstruction,
-                                  CodeReference::DataTarget,
-                                  CodeReference::MemoryDisplacementOperand,
-                                  CodeReference::Internal, 6618560);
-        }*/
-        //addr -= instruction->size();
-        if(tryEval(expr.get(), addr, a)) {
-          handleXref(cfgInstruction, a);
-          if (IsInRegion(section_manager.getRegion(".text"), a)) {
+
+        if(TryEval(expr.get(), addr, a)) {
+          HandleXref(cfg_instruction, a);
+          if (gSectionManager->IsInRegion(".text", a)) {
             // get last one and change it to code
-            auto xref = cfgInstruction->mutable_xrefs(cfgInstruction->xrefs_size() - 1);
+            auto xref = cfg_instruction->mutable_xrefs(cfg_instruction->xrefs_size() - 1);
             xref->set_operand_type(CodeReference::MemoryOperand);
           }
           direct_values[i] = a;
         }
-      //} else if (instruction_id == 8) {
       } else if (instruction->getCategory() == InstructionAPI::c_BranchInsn) {
         Address a;
-        if (tryEval(expr.get(), addr - instruction->size(), a)) {
-          auto code_ref = AddCodeXref(cfgInstruction,
-                                      CodeReference::CodeTarget,
-                                      CodeReference::ControlFlowOperand,
-                                      CodeReference::Internal, a);
-          if (isExternal(a)) {
-            code_ref->set_location(CodeReference::External);
-            code_ref->set_name(getXrefName(a));
+        // This has + |instruction| in AST
+        if (TryEval(expr.get(), addr - instruction->size(), a)) {
+          if (gDisassContext->HandleCodeXref({0, a, cfg_instruction})) {
+            auto cfg_xref = cfg_instruction->mutable_xrefs(cfg_instruction->xrefs_size() - 1);
+            cfg_xref->set_target_type(CodeReference::CodeTarget);
+            cfg_xref->set_operand_type(CodeReference::ControlFlowOperand);
           }
           direct_values[i] = a;
         }
       }
-    } else {
-      /*if (FLAGS_pie_mode &&
-          instruction->getCategory() == InstructionAPI::c_BranchInsn) {
-        //if (operands.size() == 1 ) {
-          //EXPERIMENTAL(lukas): things like jmpq *%rdx need Offset table xref
-          auto xref = cfgInstruction->add_xrefs();
-          xref->set_target_type(CodeReference::DataTarget);
-          xref->set_operand_type(CodeReference::OffsetTable);
-          xref->set_location(CodeReference::Internal);
-          xref->set_ea(4100);
-        }*/
-      //}
     }
     ++i;
-    checkDisplacement(expr.get(), cfgInstruction);
+    CheckDisplacement(expr.get(), cfg_instruction);
   }
+
+  // We may be storing some address wich is quite possibly entrypoint of
+  // something we should treat as function in cfg
   if (direct_values[0] && direct_values[1]) {
     addr -= instruction->size();
-    bool is_somewhere_reasonable = IsInRegions(
-        {
-          section_manager.getRegion(".bss"),
-          section_manager.getRegion(".data"),
-          section_manager.getRegion(".rodata"),
-        },
+    bool is_somewhere_reasonable = gSectionManager->IsInRegions(
+        {".bss", ".data", ".rodata"},
         direct_values[0]);
-    if (IsInRegion(section_manager.getRegion(".text"), direct_values[1]) &&
+    if (gSectionManager->IsInRegion(".text", direct_values[1]) &&
         is_somewhere_reasonable) {
-      LOG(INFO) << "Storing address from .text into .bss";
-      if (func_map.find(direct_values[0]) == func_map.end()) {
+      if (!gDisassContext->getInternalFunction(direct_values[0])) {
         LOG(INFO)
             << "\tAnd it is not parsed yet, storing it to be resolved later!";
         inst_xrefs_to_resolve.insert({direct_values[1], {}});
@@ -1016,17 +738,10 @@ void CFGWriter::handleNonCallInstruction(
   }
 }
 
-void CFGWriter::writeExternalFunctions() {
+void CFGWriter::WriteExternalFunctions() {
   std::vector<std::string> unknown;
-  auto known = ext_func_manager.GetAllUsed( unknown );
+  auto known = gExtFuncManager->GetAllUsed( unknown );
   LOG(INFO) << "Found " << known.size() << " external functions";
-  LOG(INFO) << "Found " << unknown.size()
-            << "possibly unknown external functions";
-
-  for (auto &name : unknown) {
-    known.push_back({name});
-    LOG(INFO) << "Possibly unknown external " << name;
-  }
 
   for (auto &func : known) {
     LOG(INFO) << "External function " << func.symbol_name;
@@ -1041,161 +756,152 @@ void CFGWriter::writeExternalFunctions() {
       }
     }
 
-    if (!found) {
-      throw std::runtime_error{"unresolved external function call"};
-    }
-    LOG(INFO) << "\tat: " << a;
-    external_functions.insert({a, func.symbol_name});
-    auto cfgExtFunc = module.add_external_funcs();
+    CHECK(found) << "External function was not found in CodeSource::linkage()";
 
-    cfgExtFunc->set_name(func.symbol_name);
-    cfgExtFunc->set_ea(a);
-    cfgExtFunc->set_cc(func.CfgCallingConvention());
-    cfgExtFunc->set_has_return(func.has_return);
-    cfgExtFunc->set_no_return(!func.has_return);
-    cfgExtFunc->set_argument_count(func.arg_count);
-    cfgExtFunc->set_is_weak(func.is_weak);
+    func.ea = a;
+    auto cfg_external_func = magic_section.WriteExternalFunction(module, func);
+    gDisassContext->external_funcs.insert({a, cfg_external_func});
   }
 }
 
-bool CFGWriter::handleDataXref(const CFGWriter::CrossXref<mcsema::Segment> &xref) {
-  return handleDataXref(xref.segment, xref.ea, xref.target_ea);
+bool CFGWriter::HandleDataXref(const CrossXref<mcsema::Segment *> &xref) {
+  return HandleDataXref(xref.segment, xref.ea, xref.target_ea);
 }
 
-bool CFGWriter::handleDataXref(mcsema::Segment *segment,
+bool CFGWriter::HandleDataXref(mcsema::Segment *segment,
                                Address ea,
                                Address target) {
-  CrossXref<mcsema::Segment> xref = {ea, target, segment};
+  CrossXref<mcsema::Segment *> context_xref = {ea, target, segment};
 
-  if (FishForXref(segment_vars ,xref)) {
-    found_xref.insert(ea);
-    return true;
-  }
-
-  if (FishForXref(func_map, xref)) {
-    found_xref.insert(ea);
-    return true;
-  }
-
-  if (FishForXref(external_vars, xref)) {
-    found_xref.insert(ea);
-    return true;
-  }
-
-  if (FishForXref(global_vars, xref)) {
-    found_xref.insert(ea);
+  // segment_vars, external_vars, global_vars
+  if (gDisassContext->HandleDataXref(context_xref)) {
     return true;
   }
   return false;
-
 }
 
 //TODO(lukas): This is finding vars actually?
-void CFGWriter::tryParseVariables(SymtabAPI::Region *region, mcsema::Segment *segment) {
+void CFGWriter::TryParseVariables(SymtabAPI::Region *region, mcsema::Segment *segment) {
   std::string base_name = region->getRegionName();
+  auto offset = static_cast<uint8_t *>(region->getPtrToRawData());
+  auto end = region->getMemOffset() + region->getMemSize();
+
   LOG(INFO) << "Trying to parse region " << region->getRegionName()
             << " for xrefs & vars";
+  LOG(INFO) << "Starts at 0x" << std::hex << region->getMemOffset()
+            << " ends at 0x" << end;
+  // Some random names
   static int counter = 0;
   static int unnamed = 0;
-  auto offset = static_cast<uint8_t *>(region->getPtrToRawData());
-  auto end = offset + region->getDiskSize();
 
   for (int j = 0; j < region->getDiskSize(); j += 1, ++offset) {
     CHECK(region->getMemOffset() + j == region->getDiskOffset() + j)
         << "Memory offset != Disk offset, investigate!";
 
+    /*
+    if ((region->getDiskSize() + j) % 4) {
+      continue;
+    }*/
+
     if (*offset == 0) {
       continue;
     }
+
+    // Read until next zero
     uint64_t size = j;
     while (*offset != 0) {
       ++j;
       ++offset;
-      if (offset + size == end) return;
+      if (region->getMemOffset() + j == end) {
+        LOG(INFO) << "Hit end of region";
+        break;
+      }
     }
 
+    // Zero was found, but it may have been something like
+    // 04 bc 00 00 so zero is still valid part of address
     uint64_t off = size % 4;
     auto diff = j - size;
     if (diff + off <= 4) {
       diff += off;
     }
-    //TODO(lukas): -fPIC & -pie ?
     if (diff <= 4 && diff  >= 3) {
 
       auto tmp_ptr = reinterpret_cast<std::uint64_t *>(static_cast<uint8_t *>(
             region->getPtrToRawData()) + size - off);
-      if (handleDataXref(segment, region->getMemOffset() + size - off, *tmp_ptr)) {
+      Dyninst::Address target_ea = *tmp_ptr;
+      Dyninst::Address ea = region->getMemOffset() + size - off;
+
+      /*if (gDisassContext->segment_vars.find(ea) !=
+          gDisassContext->segment_vars.end()) {
+        LOG(INFO) << "0x" << std::hex << ea << " is already resolved!";
+        continue;
+      }*/
+
+      if (HandleDataXref(segment, ea, target_ea)) {
         LOG(INFO) << "Fished up " << std::hex
                   << region->getMemOffset() + size << " " << *tmp_ptr;
         continue;
       }
 
-      if (IsInRegion(region, *tmp_ptr)) {
+      if (gSectionManager->IsInRegion(region, target_ea)) {
         std::string name = base_name + "_unnamed_" + std::to_string(++unnamed);
-        WriteDataXref({region->getMemOffset() + size - off,
-                      *tmp_ptr,
-                      segment},
-                      name);
+        WriteDataXref({ea, target_ea, segment}, name);
 
-        //Now add target as var
+        //Now add target as var because it was not before
         auto cfg_var = segment->add_vars();
         cfg_var->set_name(name);
-        cfg_var->set_ea(*tmp_ptr);
-        segment_vars.insert({*tmp_ptr, name});
-        found_xref.insert(region->getMemOffset() + size - off);
+        cfg_var->set_ea(target_ea);
+        gDisassContext->segment_vars.insert({target_ea, cfg_var});
         continue;
 
-      } else if (IsInBinary(code_source, *tmp_ptr)) {
-        LOG(INFO) << "Cross xref " << std::hex
-                  << region->getMemOffset() + size - off << " " << *tmp_ptr;
-        cross_xrefs.push_back({region->getMemOffset() + size,
-                              *tmp_ptr,
-                              segment});
+      } else if (gSectionManager->IsInBinary(target_ea)) {
+        LOG(INFO) << "Cross xref 0x" << std::hex
+                  << ea << " -> 0x" << target_ea;
+        cross_xrefs.push_back({ea, target_ea, segment});
         continue;
-      } else if (IsInRegion(section_manager.getRegion(".bss"), *tmp_ptr)) {
-        LOG(INFO) << "Xref into bss " << std::hex
-                  << region->getMemOffset() + size - off << " " << *tmp_ptr;
-        WriteDataXref({region->getMemOffset() + size - off,
-                      *tmp_ptr,
-                      segment});
       }
     }
     std::string name = base_name + "_" + std::to_string(counter);
     ++counter;
+    LOG(INFO) << "\tAdding var " << name << " at 0x" << std::hex << region->getMemOffset() + size;
     auto var = segment->add_vars();
     var->set_ea(region->getMemOffset() + size);
     var->set_name(name);
-    segment_vars.insert({region->getMemOffset() + size, name});
+    gDisassContext->segment_vars.insert({region->getMemOffset() + size, var});
   }
 }
 
-//TODO(lukas): Best case remove this in favor of tryParse
-void CFGWriter::xrefsInSegment(SymtabAPI::Region *region,
+void CFGWriter::XrefsInSegment(SymtabAPI::Region *region,
                                mcsema::Segment *segment) {
 
+  // Both are using something smarter, as they may contain other
+  // data as static strings
+  if (region->getRegionName() == ".data" ||
+      region->getRegionName() == ".rodata") {
+    return;
+  }
   auto offset = static_cast<std::uint64_t*>(region->getPtrToRawData());
 
   for (int j = 0; j < region->getDiskSize(); j += 8, offset++) {
-    if (found_xref.count(region->getMemOffset() + j)) {
-      LOG(WARNING) << "Dual xref detected!";
-      continue;
-    }
+    // Just so we know, there was something shady
     LOG(INFO)
       << std::hex << "Trying to resolve xref from " << region->getRegionName()
       << " at 0x" <<region->getMemOffset() + j << " targeting 0x" << *offset;
 
-    if (!handleDataXref(segment, region->getMemOffset() + j, *offset)) {
-      LOG(INFO) << "Did not resolve it, try to search in .text";
+    if (!HandleDataXref(segment, region->getMemOffset() + j, *offset)) {
+      LOG(INFO) << "\tDid not resolve it, try to search in .text";
 
-      if (IsInRegion(section_manager.getRegion(".text"), *offset)) {
-        LOG(INFO) << "Xref is pointing into .text";
-        code_xrefs_to_resolve.insert({*offset,{region->getMemOffset() + j, *offset, segment}});
+      if (gSectionManager->IsInRegion(".text", *offset)) {
+        LOG(INFO) << "\tXref is pointing into .text";
+        code_xrefs_to_resolve.insert(
+            {*offset,{region->getMemOffset() + j, *offset, segment}});
       }
     }
   }
 }
 
-void writeRawData(std::string& data, SymtabAPI::Region* region) {
+void WriteRawData(std::string& data, SymtabAPI::Region* region) {
   auto i = 0U;
 
   for (; i < region->getDiskSize(); ++i) {
@@ -1208,178 +914,170 @@ void writeRawData(std::string& data, SymtabAPI::Region* region) {
   }
 }
 
+// Writes into section on specified offset
+// If offset points beyond section, it is resized to contain it
+void WriteAsRaw(std::string& data, uint64_t number, int64_t offset) {
+  LOG(INFO) << "Writing raw " << number << " to offset " << offset;
+  if (offset < 0) {
+    LOG(FATAL) << "Trying yo Write raw on negative offset";
+  }
+  if (offset + 3 >= data.size()) {
+    LOG(WARNING) << "AsRaw would overWrite stuff";
+    data.resize(offset + 3, '\0');
+  }
+  for (int i = 0; i < 4; ++i) {
+    data[offset + i] = (number >> (i * 8));
+  }
+}
+
 //TODO(lukas): Relic of old PR, not sure if needed
 //             -fPIC & -pie?
-void CFGWriter::writeGOT(SymtabAPI::Region* region,
+void CFGWriter::WriteGOT(SymtabAPI::Region* region,
                                  mcsema::Segment* segment) {
-  auto rela_dyn = section_manager.getRegion(".rela.dyn");
+  auto rela_dyn = gSectionManager->GetRegion(".rela.dyn");
   if (!rela_dyn || !FLAGS_pie_mode) {
     return;
   }
   const auto &relocations = rela_dyn->getRelocations();
 
+  auto old_data = segment->mutable_data();
+  std::string data{*old_data};
   for (auto reloc : relocations) {
+    if (!gSectionManager->IsInRegion(region, reloc.rel_addr())) {
+      continue;
+    }
     bool found = false;
     LOG(INFO) << "Trying to resolve reloc " << reloc.name();
     for (auto ext_var : magic_section.ext_vars) {
       if (reloc.name() == ext_var->name()) {
-        LOG(INFO) << "Xref in .got io external living in magic_section";
-        auto xref = segment->add_xrefs();
-        xref->set_target_name(ext_var->name());
-        xref->set_ea(reloc.rel_addr());
-        xref->set_target_ea(ext_var->ea());
-        xref->set_target_is_code(false);
-        xref->set_target_fixup_kind(mcsema::DataReference::Absolute);
-        xref->set_width(ptr_byte_size);
+        LOG(INFO) << "Writing cfg_xref in got 0x" << std::hex
+                  << reloc.rel_addr() << " -> 0x" << ext_var->ea();
+        auto cfg_xref = segment->add_xrefs();
+        cfg_xref->set_target_name(ext_var->name());
+        cfg_xref->set_ea(reloc.rel_addr());
+        cfg_xref->set_target_ea(ext_var->ea());
+        cfg_xref->set_target_is_code(false);
+        cfg_xref->set_target_fixup_kind(mcsema::DataReference::Absolute);
+        cfg_xref->set_width(ptr_byte_size);
         found = true;
+        gDisassContext->data_xrefs.insert({reloc.rel_addr(), cfg_xref});
+        WriteAsRaw(data, ext_var->ea(), reloc.rel_addr() - segment->ea());
       }
     }
     if (!found && !reloc.name().empty()) {
       LOG(WARNING)
           << "Giving magic_space to" << reloc.name();
 
-      Address unreal_ea = magic_section.start_ea + magic_section.size;
-      //auto extVar = module.add_external_vars();
-      //extVar->set_name(reloc.name());
-      //extVar->set_ea(unreal_ea);
+      auto unreal_ea = magic_section.AllocSpace(ptr_byte_size);
+      auto cfg_xref = segment->add_xrefs();
+      cfg_xref->set_target_name(reloc.name());
+      cfg_xref->set_ea(reloc.rel_addr());
+      cfg_xref->set_target_ea(unreal_ea);
+      cfg_xref->set_target_is_code(true);
+      cfg_xref->set_target_fixup_kind(mcsema::DataReference::Absolute);
+      cfg_xref->set_width(ptr_byte_size);
+      gDisassContext->data_xrefs.insert({reloc.rel_addr(), cfg_xref});
+      WriteAsRaw(data, unreal_ea, reloc.rel_addr() - segment->ea());
 
-      //extVar->set_size(ptr_byte_size);
-
-      //TODO(lukas): This needs some checks
-      //extVar->set_is_weak(false);
-      //extVar->set_is_thread_local(false);
-
-      magic_section.size += ptr_byte_size;
-      for (int i = 0; i < ptr_byte_size; ++i) {
-        magic_section.data << "\0";
+    if (gExtFuncManager->IsExternal(reloc.name())) {
+        auto func = gExtFuncManager->GetExternalFunction(reloc.name());
+        func.imag_ea = unreal_ea;
+        func.WriteHelper(module, unreal_ea);
       }
-      //external_vars.insert( { unreal_ea, reloc.name() } );
-      //magic_section.ext_vars.push_back(extVar);
-
-      auto xref = segment->add_xrefs();
-      xref->set_target_name(reloc.name());
-      xref->set_ea(reloc.rel_addr());
-      xref->set_target_ea(unreal_ea);
-      xref->set_target_is_code(true);
-      xref->set_target_fixup_kind(mcsema::DataReference::Absolute);
-      xref->set_width(ptr_byte_size);
     }
   }
+  for (int i = 0; i < data.size(); ++i) {
+    LOG(INFO) << std::hex << +static_cast<uint8_t>((data)[i]);
+  }
+  segment->set_data(data);
 }
 
-void CFGWriter::writeRelocations(SymtabAPI::Region* region,
+void CFGWriter::WriteRelocations(SymtabAPI::Region* region,
                          mcsema::Segment *segment) {
 
-  auto rela_dyn = region;
+  auto rela_dyn = gSectionManager->GetRegion(".rela.plt");
+  if (!rela_dyn || !FLAGS_pie_mode) {
+    return;
+  }
   const auto &relocations = rela_dyn->getRelocations();
+  auto data = segment->mutable_data();
 
   for (auto reloc : relocations) {
-    for (auto f : code_object.funcs()) {
-      if (f->entry()->start() == reloc.getDynSym()->getOffset()) {
-        handleDataXref(
-            segment, reloc.rel_addr(), reloc.getDynSym()->getOffset());
-        break;
-      }
-    }
-    /*
     bool found = false;
     LOG(INFO) << "Trying to resolve reloc " << reloc.name();
-    for (auto ext_var : magic_section.ext_vars) {
-      if (reloc.name() == ext_var->name()) {
-        LOG(INFO) << "Xref in .got to external living in magic_section";
+    for (auto ext_func : magic_section.ext_funcs) {
+      if (reloc.name() == ext_func->name()) {
+        LOG(INFO) << "Writing xref in got 0x" << std::hex
+                  << reloc.rel_addr() << " -> 0x" << ext_func->ea();
         auto xref = segment->add_xrefs();
-        xref->set_target_name(ext_var->name());
+        xref->set_target_name(ext_func->name());
         xref->set_ea(reloc.rel_addr());
-        xref->set_target_ea(ext_var->ea());
-        xref->set_target_is_code(false);
+        xref->set_target_ea(ext_func->ea());
+        xref->set_target_is_code(true);
         xref->set_target_fixup_kind(mcsema::DataReference::Absolute);
         xref->set_width(ptr_byte_size);
         found = true;
+        WriteAsRaw(*data, ext_func->ea(), reloc.rel_addr() - segment->ea());
       }
     }
-    if (!found && !reloc.name().empty()) {
-      LOG(WARNING)
-          << "Giving magic_space to" << reloc.name();
-
-      Address unreal_ea = magic_section.start_ea + magic_section.size;
-      //auto extVar = module.add_external_vars();
-      //extVar->set_name(reloc.name());
-      //extVar->set_ea(unreal_ea);
-
-      //extVar->set_size(ptr_byte_size);
-
-      //TODO(lukas): This needs some checks
-      //extVar->set_is_weak(false);
-      //extVar->set_is_thread_local(false);
-
-      magic_section.size += ptr_byte_size;
-      for (int i = 0; i < ptr_byte_size; ++i) {
-        magic_section.data << "\0";
-      }
-      //external_vars.insert( { unreal_ea, reloc.name() } );
-      //magic_section.ext_vars.push_back(extVar);
-
-      auto xref = segment->add_xrefs();
-      xref->set_target_name(reloc.name());
-      xref->set_ea(reloc.rel_addr());
-      xref->set_target_ea(unreal_ea);
-      xref->set_target_is_code(true);
-      xref->set_target_fixup_kind(mcsema::DataReference::Absolute);
-      xref->set_width(ptr_byte_size);
-    }*/
   }
 }
 
 void CFGWriter::ResolveCrossXrefs() {
   for (auto &xref : cross_xrefs) {
-    auto g_var = global_vars.find(xref.target_ea);
-    if (g_var != global_vars.end()) {
+    auto g_var = gDisassContext->global_vars.find(xref.target_ea);
+    if (g_var != gDisassContext->global_vars.end()) {
       LOG(ERROR)
           << "CrossXref is targeting global variable and was not resolved earlier!";
       continue;
     }
 
-    if(!handleDataXref(xref)) {
-      LOG(WARNING) << std::hex << xref.ea << " is unresolved, targeting "
-                   << xref.target_ea;
+    if(!HandleDataXref(xref)) {
+      LOG(INFO) << std::hex << xref.ea << " is unresolved, targeting "
+                << xref.target_ea;
+      if (gSectionManager->IsInRegions({".data", ".rodata", ".bss"}, xref.target_ea)) {
+        LOG(INFO) << "It is pointing into data sections, assuming it is xref";
+        WriteDataXref(xref);
+      }
     }
     // If it's xref into .text it's highly possible it is
     // entrypoint of some function that was missed by speculative parse.
     // Let's try to parse it now
 
-    if (IsInRegion(section_manager.getText(), xref.target_ea) &&
-        IsInRegion(section_manager.getRegion(".text"), xref.target_ea)) {
-          code_xrefs_to_resolve.insert({xref.target_ea, xref});
+    if (gSectionManager->IsInRegion(".text", xref.target_ea)) {
+      LOG(INFO) << "\tIs acturally targeting something in .text!";
+      code_xrefs_to_resolve.insert({xref.target_ea, xref});
     }
   }
 }
 
-void writeBssXrefs(SymtabAPI::Region *region,
+void WriteBssXrefs(SymtabAPI::Region *region,
                    mcsema::Segment *segment,
-                   const SymbolMap &externals) {
+                   const DisassContext::SymbolMap<mcsema::ExternalVariable *> &externals) {
   for (auto &external : externals) {
-    if (IsInRegion(region, external.first)) {
+    if (gSectionManager->IsInRegion(region, external.first)) {
       auto cfg_xref = segment->add_xrefs();
       cfg_xref->set_ea(external.first);
       cfg_xref->set_width(8);
       cfg_xref->set_target_ea(external.first);
-      cfg_xref->set_target_name(external.second);
+      cfg_xref->set_target_name(external.second->name());
       cfg_xref->set_target_is_code(false);
       cfg_xref->set_target_fixup_kind(mcsema::DataReference::Absolute);
     }
   }
 }
 
-//TODO(lukas): We are writing & parsing more than is required!
-//             .rodata, .data are parsed twice!
-//             Use tryParse for all data sections
-void CFGWriter::writeInternalData() {
-  auto dataRegions = section_manager.GetDataRegions();
+// Write content of sections, try to parse for xrefs and vars
+// .rodata, .data have special treatment, since there can be string
+// variables and such
+// Write things into gDisassContext
+void CFGWriter::WriteInternalData() {
+  auto dataRegions = gSectionManager->GetDataRegions();
 
   for (auto region : dataRegions) {
+
+    // IDA does not include some of these, so neither should we
     std::set<std::string> no_parse = {
-      //".rela.dyn",
-      //".rela.plt",
       ".dynamic",
       ".dynstr",
       ".dynsym",
@@ -1398,31 +1096,31 @@ void CFGWriter::writeInternalData() {
     if ( region->getRegionName() == ".fini_array" ) {
       continue;
     }
-    auto cfgInternalData = module.add_segments();
+    auto cfg_internal_data = module.add_segments();
 
     std::string data;
-    writeRawData(data, region);
+    WriteRawData(data, region);
 
     // .init & .fini should be together
     if ( region->getRegionName() == ".init_array" ) {
       SymtabAPI::Region* fini;
       symtab.findRegion( fini, ".fini_array" );
-      writeRawData( data, fini );
-      writeDataVariables(fini, cfgInternalData);
+      WriteRawData( data, fini );
+      WriteDataVariables(fini, cfg_internal_data);
 
-      xrefsInSegment( fini, cfgInternalData );
+      XrefsInSegment( fini, cfg_internal_data );
     }
 
-    cfgInternalData->set_ea(region->getMemOffset());
-    cfgInternalData->set_data(data);
-    cfgInternalData->set_read_only(region->getRegionPermissions() ==
+    cfg_internal_data->set_ea(region->getMemOffset());
+    cfg_internal_data->set_data(data);
+    cfg_internal_data->set_read_only(region->getRegionPermissions() ==
                                    SymtabAPI::Region::RP_R);
-    cfgInternalData->set_is_external(false);
-    cfgInternalData->set_name(region->getRegionName());
-    cfgInternalData->set_is_exported(false);      /* TODO: As for now, ignored */
-    cfgInternalData->set_is_thread_local(false); /* TODO: As for now, ignored */
+    cfg_internal_data->set_is_external(false);
+    cfg_internal_data->set_name(region->getRegionName());
+    cfg_internal_data->set_is_exported(false);      /* TODO: As for now, ignored */
+    cfg_internal_data->set_is_thread_local(false); /* TODO: As for now, ignored */
 
-
+    // TODO(lukas): We probably don't need to parse these?
     std::set<std::string> dont_parse = {
       ".eh_frame",
       ".rela.dyn",
@@ -1439,40 +1137,30 @@ void CFGWriter::writeInternalData() {
       ".jcr",
     };
     if (dont_parse.find(region->getRegionName()) == dont_parse.end()) {
-      writeDataVariables(region, cfgInternalData);
+      WriteDataVariables(region, cfg_internal_data);
 
-      xrefsInSegment( region, cfgInternalData );
+      XrefsInSegment( region, cfg_internal_data );
     }
 
+    // IDA output of .bss produced some "self xrefs"
     if (region->getRegionName() == ".bss") {
-      writeBssXrefs(region, cfgInternalData, external_vars);
+      WriteBssXrefs(region, cfg_internal_data, gDisassContext->external_vars);
     }
 
+    // Apply relocations of external functions
     if (region->getRegionName() == ".got.plt") {
-      writeRelocations(region, cfgInternalData);
+      WriteRelocations(region, cfg_internal_data);
     }
 
+    // Apply relocations of external variables if needed
     if (region->getRegionName() == ".got") {
-      writeGOT(region, cfgInternalData);
-    }
-
-    //TODO(lukas): Debug print, remove on release
-    if (region->getRegionName() == ".rela.plt" ||
-        region->getRegionName() == ".rela.dyn" ||
-        region->getRegionName() == ".got") {
-      LOG(INFO) << "Dumping content of: " << region->getRegionName();
-      auto relocs = region->getRelocations();
-      for (auto &reloc : relocs) {
-        LOG(INFO) << "Entry:\n\ttarget: " << reloc.target_addr()
-                  << "\n\trel: "  << reloc.rel_addr()
-                  << "\n\taddend: " << reloc.addend();
-      }
+      WriteGOT(region, cfg_internal_data);
     }
   }
   ResolveCrossXrefs();
 }
 
-void CFGWriter::writeDataVariables(Dyninst::SymtabAPI::Region *region,
+void CFGWriter::WriteDataVariables(Dyninst::SymtabAPI::Region *region,
                                    mcsema::Segment *segment) {
   std::vector<SymtabAPI::Symbol *> vars;
   symtab.getAllSymbolsByType(vars, SymtabAPI::Symbol::ST_OBJECT);
@@ -1481,102 +1169,27 @@ void CFGWriter::writeDataVariables(Dyninst::SymtabAPI::Region *region,
     if ((a->getRegion() && a->getRegion() == region) ||
         (a->getOffset() == region->getMemOffset())) {
 
-      if (external_vars.find(a->getOffset()) != external_vars.end()) {
+      if (gDisassContext->external_vars.find(a->getOffset()) !=
+          gDisassContext->external_vars.end()) {
         continue;
       }
 
       auto var = segment->add_vars();
       var->set_ea(a->getOffset());
       var->set_name(a->getMangledName());
-
-      segment_vars.insert({a->getOffset(), a->getMangledName()});
+      LOG(INFO) << "Added var 0x" << std::hex << a->getOffset();
+      gDisassContext->segment_vars.insert({a->getOffset(), var});
     }
   }
   if (region->getRegionName() == ".rodata" ||
       region->getRegionName() == ".data") {
     LOG(INFO) << "Speculative parse of " << region->getRegionName();
-    tryParseVariables(region, segment);
+    TryParseVariables(region, segment);
   }
 }
 
 
-bool CFGWriter::isExternal(Address addr) const {
-  bool is = false;
-  if (code_object.cs()->linkage().find(addr) != code_object.cs()->linkage().end()) {
-    is = ext_func_manager.IsExternal(code_object.cs()->linkage()[addr]);
-  }
-
-  if (external_vars.find(addr) != external_vars.end()) {
-    is  = true;
-  }
-
-  return is;
-}
-
-std::string CFGWriter::getExternalName(Address addr) const {
-  auto name_hndl = code_object.cs()->linkage().find(addr);
-  if (name_hndl != code_object.cs()->linkage().end()) {
-    return name_hndl->second;
-  }
-  return "";
-}
-
-// Replace in favor of smartEval
-bool CFGWriter::tryEval(InstructionAPI::Expression *expr,
-                        const Address ip,
-                        Address &result) const {
-
-  LOG(INFO) << expr->format();
-  auto res = expr->eval();
-  LOG(INFO) << "expr->eval() " << res.format();
-  if (expr->eval().format() != "[empty]") {
-    LOG(INFO) << "Empty";
-    result = expr->eval().convert<Address>();
-    return true;
-  }
-
-  if (auto bin = dynamic_cast<InstructionAPI::BinaryFunction *>(expr)) {
-    LOG(INFO) << "BF";
-    std::vector<InstructionAPI::InstructionAST::Ptr> args;
-    bin->getChildren(args);
-
-    Address left, right;
-
-    auto first = tryEval(
-        dynamic_cast<InstructionAPI::Expression *>(args[0].get()),
-        ip, left);
-    auto second = tryEval(
-        dynamic_cast<InstructionAPI::Expression *>(args[1].get()),
-        ip, right);
-    if (first && second) {
-      if (bin->isAdd()) {
-        result = left + right;
-        return true;
-      } else if (bin->isMultiply()) {
-        result = left * right;
-        return true;
-      }
-
-      return false;
-    }
-  } else if (auto reg = dynamic_cast<InstructionAPI::RegisterAST *>(expr)) {
-    if (reg->format() == "RIP") {
-      LOG(INFO) << "RIP "<< ip;
-      result = ip;
-      return true;
-    }
-  } else if (auto imm = dynamic_cast<InstructionAPI::Immediate *>(expr)) {
-    result = imm->eval().convert<Address>();
-    LOG(INFO) << "IMM "<< result;
-    return true;
-  } else if (auto deref = dynamic_cast<InstructionAPI::Dereference *>(expr)) {
-    LOG(INFO) << "DER";
-    std::vector<InstructionAPI::InstructionAST::Ptr> args;
-    deref->getChildren(args);
-    return tryEval(dynamic_cast<InstructionAPI::Expression *>(args[0].get()),
-                   ip,
-                   result);
-  }
-  LOG(INFO) << "ERR";
-  return false;
+bool CFGWriter::IsExternal(Address addr) const {
+  return gDisassContext->external_vars.find(addr) != gDisassContext->external_vars.end() ||
+         gDisassContext->external_funcs.find(addr) != gDisassContext->external_funcs.end();
 }
