+++ tools/mcsema_disass/dyninst/ExternalFunctionManager.cpp
@@ -1,5 +1,31 @@
 #include "ExternalFunctionManager.h"
-#include <iostream>
+
+#include <glog/logging.h>
+
+std::unique_ptr<ExternalFunctionManager>
+    gExtFuncManager(new ExternalFunctionManager);
+
+mcsema::ExternalFunction *ExternalFunction::WriteHelper(
+    mcsema::Module &module,
+    uint64_t ea) {
+	auto cfg_external_func = module.add_external_funcs();
+
+	cfg_external_func->set_name(symbol_name);
+	cfg_external_func->set_ea(ea);
+	cfg_external_func->set_cc(CfgCallingConvention());
+	cfg_external_func->set_has_return(has_return);
+	cfg_external_func->set_no_return(!has_return);
+	cfg_external_func->set_argument_count(arg_count);
+	cfg_external_func->set_is_weak(is_weak);
+
+  return cfg_external_func;
+}
+
+mcsema::ExternalFunction *ExternalFunction::Write(
+    mcsema::Module &module) {
+	WriteHelper(module, ea);
+	return WriteHelper(module, imag_ea);
+}
 
 ExternalFunction::CfgCC ExternalFunction::CfgCallingConvention() const {
   switch (cc) {
@@ -18,12 +44,13 @@ void ExternalFunctionManager::AddExternalSymbol(const std::string &name,
 }
 
 void ExternalFunctionManager::AddExternalSymbol(const std::string &s) {
-  if (s.empty())
+  if (s.empty()) {
     return; // Empty line
-  else if (s.front() == '#')
+  } else if (s.front() == '#') {
     return; // Comment line
-  else if (s.substr(0, 5) == "DATA:")
+  } else if (s.substr(0, 5) == "DATA:") {
     return; // Refers to external data, not a function
+  }
 
   std::string rest = s;
   int n = rest.find(' ');
@@ -42,17 +69,16 @@ void ExternalFunctionManager::AddExternalSymbol(const std::string &s) {
         char cc = rest.front();
         ExternalFunction::CallingConvention callConv;
 
-        if (cc == 'C')
+        if (cc == 'C') {
           callConv = ExternalFunction::CallingConvention::CallerCleanup;
-        else if (cc == 'E')
+        } else if (cc == 'E') {
           callConv = ExternalFunction::CallingConvention::CalleeCleanup;
-        else if (cc == 'F')
+        } else if (cc == 'F') {
           callConv = ExternalFunction::CallingConvention::FastCall;
-        else {
-          std::cerr << "Error while parsing symbol definition \"" << s
-                    << "\": unknown calling convention '" << cc << "'"
-                    << std::endl;
-          throw std::runtime_error{"error while parsing symbol definition"};
+        } else {
+          LOG(FATAL) << "Error while parsing symbol definition \"" << s
+                     << "\": unknown calling convention '" << cc << "'"
+                     << std::endl;
         }
 
         rest = rest.substr(n + 1);
@@ -63,47 +89,41 @@ void ExternalFunctionManager::AddExternalSymbol(const std::string &s) {
         if (n != std::string::npos) {
           ret = rest.substr(0, 1);
           signature = rest.substr(2);
-        } else
+        } else {
           ret = rest;
+        }
 
         bool noReturn;
-        if (ret == "N")
+        if (ret == "N") {
           noReturn = false;
-        else if (ret == "Y")
+        } else if (ret == "Y") {
           noReturn = true;
-        else {
-          std::cerr << "Error while parsing symbol definition \"" << s
-                    << "\": unknown return type specifier '" << noReturn << "'"
-                    << std::endl;
-          throw std::runtime_error{"error while parsing symbol definition"};
+        } else {
+          LOG(FATAL) << "Error while parsing symbol definition \"" << s
+                     << "\": unknown return type specifier '" << noReturn << "'"
+                     << std::endl;
         }
 
         bool is_weak = true;
-        if (symbolName == "__gmon_start__") {
-          is_weak = true;
-        }
+
         ExternalFunction func{symbolName, callConv, !noReturn, argCount,
                               is_weak, signature};
 
-        external_funcs[symbolName] = func;
+        external_funcs[symbolName] = std::move(func);
         return;
       }
     } else {
-      throw std::runtime_error{
-          "internal McSema call convention is no longer supported"};
+      LOG(FATAL) << "Internal McSema call convention is no longer supported";
     }
   }
 
-  std::cerr << "Error while parsing symbol definition \"" << s
-            << "\": ill-formed symbol definition" << std::endl;
-
-  throw std::runtime_error{"error while parsing symbol definition"};
+  LOG(FATAL) << "Error while parsing symbol definition \"" << s
+             << "\": ill-formed symbol definition" << std::endl;
 }
 
 void ExternalFunctionManager::AddExternalSymbols(std::istream &s) {
-  while (!s.eof()) {
-    std::string line;
-    std::getline(s, line);
+  std::string line;
+  while (std::getline(s, line)) {
     AddExternalSymbol(line);
   }
 }
@@ -117,9 +137,12 @@ bool ExternalFunctionManager::IsExternal(const std::string &name) const {
   return external_funcs.find(name) != external_funcs.end();
 }
 
-const ExternalFunction &
+ExternalFunction &
 ExternalFunctionManager::GetExternalFunction(const std::string &name) {
-  return external_funcs.at(name);
+  auto external_func = external_funcs.find(name);
+  CHECK(external_func != external_funcs.end())
+      << "External function " << name << " not found in manager";
+  return external_func->second;
 }
 
 void ExternalFunctionManager::ClearUsed() { used_funcs.clear(); }
@@ -134,11 +157,15 @@ std::vector<ExternalFunction> ExternalFunctionManager::GetAllUsed(
   std::vector<ExternalFunction> result;
 
   for (auto name : used_funcs) {
-    try {
-      result.push_back(external_funcs.at(name));
-    } catch (const std::out_of_range &oor) {
-      unknowns.push_back(name);
+    auto external_func = external_funcs.find(name);
+    if (external_func == external_funcs.end()) {
+      LOG(INFO) << "External function " << name
+                << " not found in file with external definitions";
+      result.push_back({name});
+      continue;
     }
+    result.push_back(external_funcs.at(name));
+    unknowns.push_back(name);
   }
 
   return result;
