+++ tools/mcsema_disass/dyninst/tests/exec_tests.sh
@@ -2,7 +2,10 @@
 
 build_path=$1
 
-./run_tests.py --disass $build_path/tools/mcsema_disass/dyninst/mcsema-dyninst-disass --std_defs $build_path/../../../tools/mcsema/tools/mcsema_disass/defs/linux.txt --lift $build_path/mcsema-lift-4.0 --lib_dir $build_path/mcsema/Arch/X86/Runtime/ > /dev/null
+./run_tests.py --disass $build_path/tools/mcsema_disass/dyninst_devel/mcsema-dyninst-devel --std_defs $build_path/../../../tools/mcsema/tools/mcsema_disass/defs/linux.txt --lift $build_path/mcsema-lift-4.0 --lib_dir $build_path/mcsema/Arch/X86/Runtime/ > /dev/null
 
-./no_flags_tests.py --disass $build_path/tools/mcsema_disass/dyninst/mcsema-dyninst-disass --std_defs $build_path/../../../tools/mcsema/tools/mcsema_disass/defs/linux.txt --lift $build_path/mcsema-lift-4.0 --lib_dir $build_path/mcsema/Arch/X86/Runtime/ > /dev/null
+./no_flags_tests.py --disass $build_path/tools/mcsema_disass/dyninst_devel/mcsema-dyninst-devel --std_defs $build_path/../../../tools/mcsema/tools/mcsema_disass/defs/linux.txt --lift $build_path/mcsema-lift-4.0 --lib_dir $build_path/mcsema/Arch/X86/Runtime/ > /dev/null
 
+./pie_tests.py --disass $build_path/tools/mcsema_disass/dyninst_devel/mcsema-dyninst-devel --std_defs $build_path/../../../tools/mcsema/tools/mcsema_disass/defs/linux.txt --lift $build_path/mcsema-lift-4.0 --lib_dir $build_path/mcsema/Arch/X86/Runtime/ > /dev/null
+
+./pie_s_tests.py --disass $build_path/tools/mcsema_disass/dyninst_devel/mcsema-dyninst-devel --std_defs $build_path/../../../tools/mcsema/tools/mcsema_disass/defs/linux.txt --lift $build_path/mcsema-lift-4.0 --lib_dir $build_path/mcsema/Arch/X86/Runtime/ > /dev/null
