+++ tools/mcsema_disass/dyninst/tests/output.txt
@@ -1,0 +1,211 @@
+digraph G {
+	 subgraph cluster_0 { 
+		 label="_init"; 
+		 color=blue;
+		"4004b8" [shape=box]
+		"4004b8" [label = "_init\n4004b8"];
+		"4004b8";
+		"4004c8";
+		"4004cd";
+	}
+	"4004b8" -> "4004cd"
+	"4004b8" -> "4004c8"
+	"4004c8" -> "400520" [color=blue]
+	"4004c8" -> "4004cd"
+	"4004cd" -> "ffffffffffffffff" [color=green]
+	"4004cd" -> "400741" [color=green]
+
+	 subgraph cluster_0 { 
+		 label="write"; 
+		 color=blue;
+		"4004f0" [shape=box]
+		"4004f0" [label = "write\n4004f0"];
+		"4004f0";
+	}
+	"4004f0" -> "ffffffffffffffff"
+
+	 subgraph cluster_0 { 
+		 label="printf"; 
+		 color=blue;
+		"400500" [shape=box]
+		"400500" [label = "printf\n400500"];
+		"400500";
+	}
+	"400500" -> "ffffffffffffffff"
+
+	 subgraph cluster_0 { 
+		 label="__libc_start_main"; 
+		 color=blue;
+		"400510" [shape=box]
+		"400510" [label = "__libc_start_main\n400510"];
+		"400510";
+	}
+	"400510" -> "ffffffffffffffff"
+
+	 subgraph cluster_0 { 
+		 label="targ400520"; 
+		 color=blue;
+		"400520" [shape=box]
+		"400520" [label = "targ400520\n400520"];
+		"400520";
+	}
+	"400520" -> "ffffffffffffffff"
+
+	 subgraph cluster_0 { 
+		 label="_start"; 
+		 color=blue;
+		"400530" [shape=box,color=red]
+		"400530" [label = "_start\n400530"];
+		"400530";
+		"400559";
+	}
+	"400530" -> "400510" [color=blue]
+	"400530" -> "400559"
+
+	 subgraph cluster_0 { 
+		 label="deregister_tm_clones"; 
+		 color=blue;
+		"400560" [shape=box]
+		"400560" [label = "deregister_tm_clones\n400560"];
+		"400560";
+		"400575";
+		"40057f";
+		"400590";
+	}
+	"400560" -> "400590"
+	"400560" -> "400575"
+	"400575" -> "400590"
+	"400575" -> "40057f"
+	"40057f" -> "ffffffffffffffff"
+	"400590" -> "ffffffffffffffff" [color=green]
+
+	 subgraph cluster_0 { 
+		 label="register_tm_clones"; 
+		 color=blue;
+		"4005a0" [shape=box]
+		"4005a0" [label = "register_tm_clones\n4005a0"];
+		"4005a0";
+		"4005c3";
+		"4005cd";
+		"4005d8";
+	}
+	"4005a0" -> "4005d8"
+	"4005a0" -> "4005c3"
+	"4005c3" -> "4005d8"
+	"4005c3" -> "4005cd"
+	"4005cd" -> "ffffffffffffffff"
+	"4005d8" -> "ffffffffffffffff" [color=green]
+
+	 subgraph cluster_0 { 
+		 label="__do_global_dtors_aux"; 
+		 color=blue;
+		"4005e0" [shape=box]
+		"4005e0" [label = "__do_global_dtors_aux\n4005e0"];
+		"4005e0";
+		"4005e9";
+		"4005f2";
+		"4005fa";
+	}
+	"4005e0" -> "4005fa"
+	"4005e0" -> "4005e9"
+	"4005e9" -> "400560" [color=blue]
+	"4005e9" -> "4005f2"
+	"4005f2" -> "4005fa"
+	"4005fa" -> "ffffffffffffffff" [color=green]
+
+	 subgraph cluster_0 { 
+		 label="frame_dummy"; 
+		 color=blue;
+		"400600" [shape=box]
+		"400600" [label = "frame_dummy\n400600"];
+		"400600";
+		"40060b";
+		"400610";
+		"40061a";
+		"400620";
+	}
+	"400600" -> "400610"
+	"400600" -> "40060b"
+	"40060b" -> "4005a0"
+	"400610" -> "40060b"
+	"400610" -> "40061a"
+	"40061a" -> "ffffffffffffffff" [color=blue]
+	"40061a" -> "400620"
+	"400620" -> "4005a0"
+
+	 subgraph cluster_0 { 
+		 label="writeit"; 
+		 color=blue;
+		"400630" [shape=box]
+		"400630" [label = "writeit\n400630"];
+		"400630";
+		"400653";
+		"400686";
+	}
+	"400630" -> "4004f0" [color=blue]
+	"400630" -> "400653"
+	"400653" -> "4004f0" [color=blue]
+	"400653" -> "400686"
+	"400686" -> "ffffffffffffffff" [color=green]
+	"400686" -> "4006bc" [color=green]
+
+	 subgraph cluster_0 { 
+		 label="main"; 
+		 color=blue;
+		"4006a0" [shape=box]
+		"4006a0" [label = "main\n4006a0"];
+		"4006a0";
+		"4006bc";
+		"4006db";
+		"4006fd";
+	}
+	"4006a0" -> "400630" [color=blue]
+	"4006a0" -> "4006bc"
+	"4006bc" -> "400500" [color=blue]
+	"4006bc" -> "4006db"
+	"4006db" -> "400500" [color=blue]
+	"4006db" -> "4006fd"
+	"4006fd" -> "ffffffffffffffff" [color=green]
+
+	 subgraph cluster_0 { 
+		 label="__libc_csu_init"; 
+		 color=blue;
+		"400710" [shape=box]
+		"400710" [label = "__libc_csu_init\n400710"];
+		"400710";
+		"400741";
+		"400746";
+		"400750";
+		"40075d";
+		"400766";
+	}
+	"400710" -> "4004b8" [color=blue]
+	"400710" -> "400741"
+	"400741" -> "400766"
+	"400741" -> "400746"
+	"400746" -> "400750"
+	"400750" -> "ffffffffffffffff" [color=blue]
+	"400750" -> "40075d"
+	"40075d" -> "400750"
+	"40075d" -> "400766"
+	"400766" -> "ffffffffffffffff" [color=green]
+
+	 subgraph cluster_0 { 
+		 label="__libc_csu_fini"; 
+		 color=blue;
+		"400780" [shape=box]
+		"400780" [label = "__libc_csu_fini\n400780"];
+		"400780";
+	}
+	"400780" -> "ffffffffffffffff" [color=green]
+
+	 subgraph cluster_0 { 
+		 label="_fini"; 
+		 color=blue;
+		"400784" [shape=box]
+		"400784" [label = "_fini\n400784"];
+		"400784";
+	}
+	"400784" -> "ffffffffffffffff" [color=green]
+
+}
